moved modules/loss dir into modules/rlhf/loss and updated references
